+++ drake/examples/bouncing_ball/ball.cc
@@ -1,4 +1,4 @@
-#include "drake/examples/BouncingBall/ball-inl.h"
+#include "drake/examples/bouncing_ball/ball-inl.h"
 
 #include "drake/common/eigen_autodiff_types.h"
 #include "drake/drakeBouncingBall_export.h"
