+++ drake/examples/bouncing_ball/bouncing_ball.cc
@@ -1,4 +1,4 @@
-#include "drake/examples/BouncingBall/bouncing_ball-inl.h"
+#include "drake/examples/bouncing_ball/bouncing_ball-inl.h"
 
 #include "drake/common/eigen_autodiff_types.h"
 #include "drake/drakeBouncingBall_export.h"
